pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["paullosergio <paulomg1996@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
groq = "^0.11.0"
langchain = "^0.3.7"
langchain-groq = "^0.2.1"
python-decouple = "^3.8"
flask = "^3.0.3"
streamlit = "^1.39.0"
chromadb = "^0.5.17"
protobuf = "3.20.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
